Conflicts:
	Core/Code/Interactions/mitkMouseModeSwitcher.cpp
	Core/Code/Rendering/mitkImageVtkMapper2D.cpp

+++ Modules/IGT/IGTFilters/mitkNavigationDataTransformFilter.h
@@ -37,7 +37,7 @@ namespace mitk {
 
   public:
 
-    typedef itk::Rigid3DTransform< float > TransformType;
+    typedef itk::Rigid3DTransform< double > TransformType;
 
     mitkClassMacro(NavigationDataTransformFilter, NavigationDataToNavigationDataFilter);
     itkNewMacro(Self);
